requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Flask==1.0.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-error-reporting==0.32.0
google-cloud-firestore==1.6.0
google-cloud-logging==1.12.1
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.23.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
parsel==1.5.2
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
rsa==4.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
typed-ast==1.4.0
urllib3==1.25.3
w3lib==1.21.0
Werkzeug==0.15.5
wrapt==1.11.2
zope.interface==4.6.0
